Add tests for management command.
